+++ packages/app-page-builder/src/editor/plugins/elements/cell/Cell.tsx
@@ -1,29 +1,14 @@
 import React from "react";
-import styled from "@emotion/styled";
-import DropZone from "../../../components/DropZone";
-import Element from "../../../components/Element";
-import { DragObjectWithTypeWithTarget } from "../../../components/Droppable";
+import { PbEditorElement } from "~/types";
+import CellContainer from "./CellContainer";
 
-const CellStyle = styled("div")({
-    position: "relative"
-});
-interface CellPropsType {
-    id: string;
-    dropElement: (source: DragObjectWithTypeWithTarget, index: number) => void;
-    type: string;
-    index: number;
-    isLast?: boolean;
+interface CellProps {
+    element: PbEditorElement;
+    isActive: boolean;
 }
-const Cell: React.FC<CellPropsType> = ({ id, dropElement, index, isLast = false, type }) => {
-    return (
-        <CellStyle>
-            <DropZone.Above type={type} onDrop={source => dropElement(source, index)} />
-            <Element id={id} />
-            {isLast && (
-                <DropZone.Below type={type} onDrop={source => dropElement(source, index + 1)} />
-            )}
-        </CellStyle>
-    );
+
+const Cell: React.FC<CellProps> = props => {
+    return <CellContainer {...props} elementId={props.element.id} />;
 };
 
-export default React.memo(Cell);
+export default Cell;
